split tests into models and views
